feat: implement user authentication and authorization with JWT, add user and contact management routes

+++ my-contacts-backend/controllers/contactsController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 //@desc Get a single contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -26,7 +26,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Create a new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     console.log(req.body);
 
@@ -36,20 +36,26 @@ const createContact = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error("All fields are mandatory!");
     }
-    const contact = await Contact.create({ name, email, phone });
+    const contact = await Contact.create({ name, email, phone, user_id: req.user.id });
 
     res.status(201).json({ message: "New contact created!", contactInfo: contact });
 });
 
 //@desc Update a contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to update other user's contact");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -63,12 +69,16 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete a contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
      const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User don't have permission to delete other user's contact");
     }
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json({ message: `Contact with id: ${req.params.id} deleted!` });
